Overloaded interp to allow storing common kdtree
One instance of interp now takes property object and kdtree object
as inputs to generate interpolation function. kdtree can be
created for a space using get_kdtree function and can be used for
several properties on the same space. See test_interp function
for example of how this case is used. Eventually, we should make an
option of storing a kdtree in space object in IMAS data structure.