project_prop_on_subset no returns grid centers
project_prop_on_subset! now returns a tuple of two objects, first is
an array of tuples representing (r,z) coordinates where the values
are computed and the second is an array of the property values.

The return tuple is different if projection is done within same subset
dimensions. In this case, space keyword is not required. The first
element in the return tuple is an array of indices of the subset
objects where the value is defined. Second element is the same, the
property values that were just computed and stored.

For utility, get_subset_centers function has been added to quickly
get an array of tuples representing (r, z) coordinates of the cell
centers or the edge centers of the subset.