Adding interp(::OMAS.equilibirum__time_slice)
Brought over equilibrium 2d profile (r,z) to 1d profile (rho)
interpolation from core_profile_2d funciton in SD4SOLPS
https://github.com/ProjectTorreyPines/SD4SOLPS.jl/blob/99a36dddbfc2bdccf5a7e5479e61e9cf87bc62c9/src/SD4SOLPS.jl#L149

Example use:
rz2rho = interp(dd.equilibrium.time_slice[1])
rho = rz2rho.([(r, z) for r in 3:0.01:9, for z in -5:0.01:5])

Also added argument parsing in the test for running individual tests.
Note, that a test for testing this method of interp has not been added
yet as it will involve requiring the SD4SOLPS.geqdsk_to_imas function.
We can not make Sd4SOLPS a dependency of GGDUtils as the dependency
flow is the other way round.
I still have to figure out how to get isolated environment in test
so that this can be done. But I tested this function separately in a
scratch space.
This also suggests that geqdsk_to_imas can be moved to a more basic
repo like SOLPS2IMAS although this package is supposed to be specialized
for SOLPS only. So maybe a new I/O package?