Added option for reusing TPS matrices
While doing thin plate spline method for interpolation, it is beneficial
to reuse the matrix calculations done if the same grid susbset is
being used. So just like sharing a KDTree object, we can share the
TPS_mats which is a tuple of matrices and vectors required for this
spline function.