Added back OMAS in makefile
OMAS was replaced with IMASDD in makefile but since the currently we
are simply importing OMAS as IMASDD, we need to add back OMAS in the
makefile.

I have left IMASDD in the makefile as well, so that we can easily switch
between the two in future.

I've tested this make version on my machine.