Installing CompatHelper workflow
Installing compat helper to keep compatibility bounded to the latest
version of the dependencies.

For now, the PRs are made to this branch to get the first version of the
Project.toml file with compat entries. Once the compat entries are added,
the PRs will be made to the dev branch where if the tests pass, they
will be added and nerge with main branch will happen with version
increment.