Name change: GGDUtils -> IMASggd
We are changing the name of this package from GGDUtils to IMASggd. This
is to conform with naming requirements of Julia Registry and easilty identify
this package as part of the largest IMAS family of packages. No new functionality
has been added in this commit. Only GGDUtils has been replaced with IMASggd whereever
it appears.