Require text JuliaRegistries/TagBot in tag_bot.yml
JuliaTagBot requires one of the workflow files to have text
JuliaRegistries/TagBot to makesure it is being called. Since we are
using reusable workflows, this text is not mentioned in the calling
action file. Thus added a comment that has this text in it.

See https://github.com/JuliaRegistries/TagBot/issues/382#issuecomment-2616980328
for more details.